combine two eval modules
